added new network

+++ src/data_preprocessing.py
@@ -28,7 +28,7 @@ class Dataset:
         # model can understand better
         labelencoder_Y = LabelEncoder()
         self._Y = labelencoder_Y.fit_transform(self._Y)
-        self._Y = np.expand_dims(self._Y, 1)
+        self._Y = np.expand_dims(self._Y, 1).astype(np.float32)
 
     def create_train_test_dataset(self):
         X_train, X_test, y_train, y_test = train_test_split(self._X, self._Y, test_size=0.2,
@@ -50,7 +50,6 @@ class BCWDataset(Dataset):
         X = X.astype(np.float32)
         self._X = X
         self._Y = Y
-
 
         print("breast cancer wisconsin cancer dataset dimensions : {}".format(data.shape))
         super(BCWDataset, self).process_dataset()
